+++ srcs/init/init.c
@@ -1,5 +1,12 @@
 #include "minishell.h"
 
+// SIGINTとSIGQUITに対してシグナルハンドラなどの設定
+void	init_signals(void)
+{
+	set_main_sigint();
+	set_main_sigquit();
+}
+
 char	**get_envp_copy(char **envp)
 {
 	char	**cp;
