--- srcs/init/init_signal.c
@@ -1,8 +1,0 @@
-#include "minishell.h"
-
-// SIGINTとSIGQUITに対してシグナルハンドラなどの設定
-void	init_signals(void)
-{
-	set_main_sigint();
-	set_main_sigquit();
-}
